dev-requirements.txt
repoze.profile ==2.1
Sphinx >=2.1.2,<2.2
pytest >=5.0.1,<6
python-magic >=0.4.6,<0.5
pyprof2calltree >=1.3.2,<1.4
colorlog >=2.6,<2.7
sphinx_rtd_theme >=0.4,<0.5
